requirements.txt
asgiref==3.7.2
cloudinary==1.33.0
crispy-bootstrap5==0.7
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.19
django-allauth==0.54.0
django-crispy-forms==2.0
gunicorn==20.1.0
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==1.26.16
